introduce factory method for scalar float tensor

+++ src/test/groovy/ut/optimization/ADAM_Spec.groovy
@@ -13,7 +13,7 @@ import spock.lang.Subject
 @Subject([Optimizer])
 class ADAM_Spec extends Specification
 {
-    @Shared Tsr<?> w = Tsr.of(0)
+    @Shared Tsr<?> w = Tsr.of(0d)
     @Shared Optimizer<?> o = new ADAM<>(w)
 
     def setupSpec()
@@ -54,7 +54,7 @@ class ADAM_Spec extends Specification
         given : 'A new scalar gradient tensor is being created.'
             Tsr g = Tsr.of(expectedWeight)
         and : 'The following input is being applied to the tensor (and internal optimizer)...'
-            w.set( Tsr.of( gradient ) )
+            w.set( Tsr.of( (double)gradient ) )
             w.applyGradient()
 
         expect : 'The following state emerges:'
